Model execution bugfix, added missing requirements

+++ pbpk/openLoopSimulator.py
@@ -14,9 +14,9 @@ def input_profile(tf, tlist, ulist):
     inp = np.zeros((t.shape[0]))
     n = len(tlist) - 1
     m = len(t) -1
-    inp[0:round(m * tlist[1]/tf)] = ulist[0]
+    inp[0:int(round(m * tlist[1]/tf))] = ulist[0]
     for i in range(1, n-1): ### Edw htan range(1, n)
-        inp[round(m * tlist[i]/tf) : round(m * tlist[i+1]/tf) + 1] = ulist[i]
+        inp[int(round(m * tlist[i]/tf)) : int(round(m * tlist[i+1]/tf) + 1)] = ulist[i]
     return t, inp
 
 class SimulatorOpenLoop:
@@ -35,9 +35,9 @@ class SimulatorOpenLoop:
         inp = np.zeros((t.shape[0]))
         n = len(tlist) - 1
         m = len(t) - 1
-        inp[0:round(m * tlist[1]/T)] = ulist[0]
+        inp[0:int(round(m * tlist[1]/T))] = ulist[0]
         for i in range(1, n-1):  # Before: range(1, n)
-            inp[round(m * tlist[i]/T): round(m * tlist[i+1]/T) + 1] = ulist[i]
+            inp[int(round(m * tlist[i]/T)): int(round(m * tlist[i+1]/T) + 1)] = ulist[i]
         return t, inp
 
     def Simulate(self):
